# docs/api/OPENAPI_SPECIFICATION.yaml
openapi: 3.0.3
info:
  title: PsychSync API
  description: |
    PsychSync is a psychological assessment SaaS platform that provides personality assessments,
    team analytics, and AI-powered insights for organizations.

    **Authentication:**
    - All endpoints require JWT authentication unless marked as `public`
    - Include token in Authorization header: `Bearer <token>`
    - Token refresh endpoint available for obtaining new access tokens

    **Rate Limiting:**
    - Rate limits vary by endpoint (see specific endpoint documentation)
    - Rate limit headers included in all responses:
      - `X-RateLimit-Limit`: Max requests per window
      - `X-RateLimit-Remaining`: Remaining requests
      - `X-RateLimit-Reset`: Reset time (Unix timestamp)

    **Response Format:**
    - Success responses follow the schema defined in each endpoint
    - Error responses follow standard error format

    **Support:**
    - API Documentation: https://docs.psychsync.com/api
    - Support Email: api-support@psychsync.com
    - Status Page: https://status.psychsync.com
  version: 1.0.0
  contact:
    name: PsychSync API Team
    email: api-support@psychsync.com
    url: https://www.psychsync.com
  license:
    name: Proprietary
    url: https://www.psychsync.com/terms
  termsOfService: https://www.psychsync.com/terms
servers:
  - url: https://api.psychsync.com/api/v1
    description: Production server
  - url: https://api-staging.psychsync.com/api/v1
    description: Staging server
...
paths:
...
  /assessments/{id}:
    get:
      tags:
        - Assessments
      summary: Get assessment details
      description: Retrieve detailed information about a specific assessment including questions and sections.
      operationId: getAssessment
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Assessment UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Assessment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentDetail'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Assessment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Assessments
      summary: Update assessment
      description: Update an existing assessment. User must be the assessment creator or have admin permissions.
      operationId: updateAssessment
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Assessment UUID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssessmentUpdate'
      responses:
        '200':
          description: Assessment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assessment'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Assessment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Assessments
      summary: Delete assessment
      description: Delete an assessment. User must be the assessment creator or have admin permissions.
      operationId: deleteAssessment
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Assessment UUID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Assessment deleted successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Assessment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /teams/{id}:
    get:
      tags:
        - Teams
      summary: Get team details
      description: Retrieve detailed information about a specific team.
      operationId: getTeam
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Team UUID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Team details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamDetail'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - not team member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Team not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/token-fixed endpoint
  schemas:
    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Human-readable error message
        error_code:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
        status_code:
          type: integer
          description: HTTP status code
          example: 400
        request_id:
          type: string
          format: uuid
          description: Unique request ID for troubleshooting
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          type: string
          enum: [USER, TEAM_LEAD, ADMIN]
        is_active:
          type: boolean
        is_verified:
          type: boolean
        two_factor_enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Assessment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [PERSONALITY, SKILLS, BEHAVIORAL, COGNITIVE]
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
        framework_code:
          type: string
          example: BIG_FIVE
        created_by_id:
          type: string
          format: uuid
        team_id:
          type: string
          format: uuid
          nullable: true
        organization_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AssessmentUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
    AssessmentDetail:
      allOf:
        - $ref: '#/components/schemas/Assessment'
        - type: object
          properties:
            sections:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  title:
                    type: string
                  description:
                    type: string
                  order:
                    type: integer
                  questions:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        question_type:
                          type: string
                        question_text:
                          type: string
                        order:
                          type: integer
                        is_required:
                          type: boolean
                        config:
                          type: object
...
    Team:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        organization_id:
          type: string
          format: uuid
        created_by_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TeamDetail:
      allOf:
        - $ref: '#/components/schemas/Team'
        - type: object
          properties:
            members:
              type: array
              items:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  role:
                    type: string
                    enum: [OWNER, ADMIN, MEMBER]
                  joined_at:
                    type: string
                    format: date-time